Unsuccedeed java script side support for protocol buffers.
Untested support for protocol buffers is added in the java script at the subscription of "updateSendPartialComplete" where we expected the binary data to come as params.response. The binary blob containing the protobuf never entered this piece of code, even though it is requested successfully from the browser.

I have not been able to find out where our protobuf blob flows through the java script, and we now decide to tackle the problem from a different angle. This is therefore likely to be the last commit on this branch, and the branch will be abandoned.